Rename `databricks.sdk.errors.mapping` to `databricks.sdk.errors.platform`
`databricks.sdk.errors.mapping.ResourceDoesNotExist: Path (/Users/serge.smertin@databricks.com/.vMj0/state.json) doesn't exist.` is not as clear as `databricks.sdk.errors.platform.ResourceDoesNotExist: ...`, where it's obvious errors are coming from the Databricks Platform.
