+++ databricks/sdk/errors/__init__.py
@@ -1,4 +1,4 @@
 from .base import DatabricksError, ErrorDetail
 from .mapper import error_mapper
-from .mapping import *
+from .platform import *
 from .sdk import *
